+++ src/MySqlCdc/Providers/MySql/Metadata/TableMetadata.cs
@@ -186,7 +186,7 @@ namespace MySqlCdc.Providers.MySql
             var bytesNumber = (bitsNumber + 7) / 8;
             for (int i = 0; i < bytesNumber; i++)
             {
-                byte value = (byte)reader.ReadByte();
+                byte value = reader.ReadByte();
                 for (int y = 0; y < 8; y++)
                 {
                     int index = (i << 3) + y;
